pyproject.toml
[project]
name = "todo-console-app"
version = "0.1.0"
description = "A Python console application for managing todo tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.8.0"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"